adds same domain redirects as safe, adds comments, cleans up code
