+++ package.json
@@ -7,24 +7,29 @@
         "build": "vite build",
         "preview": "vite preview",
         "check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
-        "check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch"
+        "check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
+        "broken-link-check": "ts-node-esm scripts/check_urls.ts"
     },
     "devDependencies": {
         "@sveltejs/adapter-auto": "^1.0.0",
         "@sveltejs/adapter-static": "^1.0.0",
         "@sveltejs/kit": "^1.0.0",
+        "@types/lodash.partition": "^4.6.7",
         "autoprefixer": "^10.4.13",
+        "lodash.partition": "^4.6.0",
         "postcss": "^8.4.20",
         "svelte": "^3.54.0",
         "svelte-check": "^2.9.2",
         "tailwindcss": "^3.2.4",
         "tslib": "^2.4.1",
         "typescript": "^4.9.3",
-        "vite": "^4.0.0"
+        "vite": "^4.0.0",
+        "yaml": "^2.2.0",
+        "tldts": "^5.7.112"
     },
     "type": "module",
     "dependencies": {
-        "vitest": "^0.26.2",
-        "yaml": "^2.2.0"
+        "ts-node": "^10.4.0",
+        "vitest": "^0.26.2"
     }
 }
