--- scripts/check_urls/.gitignore
@@ -1,1 +1,0 @@
-check_urls.js
